requirements.txt
# Gmail API
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# LangChain and AI
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.0.0

# Vector Store
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.0.0

# Environment Variables
python-dotenv>=1.0.0
